Very basic command line interpreter
Main goal was to get to a way of reliaby starting the DFU bootloader.
I pulled the CommandProcessor loop and MultiTask from the other
ToasterControl repo. Refactored CommandProcessor to abstract it from
the parser. Not sure if this abstraction is valuable but it does the job
for now.

The main aim is to:
- create an easy way to add commands
- have reusable code.

Most code bases simply use a switch statement, which is more efficient
and probably suites embedded firmware better. I feel like we should use
the C++ features though. Maybe thats a bad idea.

Occurs to me that there is quite a bit of overhead to this scheme.

What's the pay off? Well for a start the serial handling code is decoupled
from the commands themselves.

It feels to complex though. Anyway - for now it works we can simplify
later.