Remove all the abstraction from CommandProcessor
I wasn't happy with the abstraction - feels like its premature to apply
this much abstraction - without it the code is simpler.

It would be nice to have something a bit more reusable so i've left the
old attempt in a seperate branch. Might come back to it. But likely
with a simpler implementation?