Rename ReadSampler to PinSampler
PinSampler is way more descriptive - there is nothing about
this class that makes it specific to the read line on the floppy.

Sampler perhaps is still a bit misleading since we're returning the
time between transitions rather an fixed interval sampling.

Not sure what the name for this should be though.