Fix interrupt based pin PinSampler
Doh! *this was stopping the callbacks from working.

Was this dereferencng this and making a copy? Not sure C++ too rusty.

The sampling seems to work but the buffer overflows on anything
high frequency.

If we get a sample every 2ms for 166ms then thats actually a buffer size
of only 133 entries - so the next step is to start the sampler when we
see the index hole and stop it once we see in spin around again.

I wonder if we should sample multiple rotations?