Attempt to fix interrupt based sampler
1. Do't use RingBug - attempt reduce ISR overhead
2. Capture all the state variables - useful to debug
3. Don't try deal with rollovers via update ISR

This one is interesting - we just change the code to add 65535 if the
input capture is less than the last capture. This means there is no
way to tell if the capture has looped more than once.

The reference manual for STM32 however tthere is hardware flag for
over capture to signal this condition - probably better to use this.

I think the rollover counter saw racing with input capture ISR so that
implementation is a bad idea.

Would neat to detect very long samples but for our purposes I don't
think this will happen in practice. Maybe use a 32bit counter if interesting
does. Or maybe the repetition counter could be used to detect these.
e.g. every 2 repitions record a value (could have race conditions again
though?)

Tried to get it to work consistently with a 3us period but this seems
to hang the CPU. I assume that the ISR code is taking so lon that it
swamping the processor and it can't handle the load.

This suggests we have to abandon the interrupt based sampling to get
the time base the FDD read line requires 3us precision.

So next attempt will be using the DMA ring buffer