Add alternate to using HardwareTimer
If we use the HardwareTimer class we can't trap the DMA interrupt
because HAL_TIM_IC_CaptureCallback is defined as part of that
library.

So instead we implement starting the DMA ourselves in order to
set the DMA complete handler.

This is a bit ugly because we're relying on the HAL internals to
remain the same.

So the other option is to use HAL directly ourselves.

But I haven't got that working yet so hence the switch.