Increase data throughput to avoid overflow error
With a 3us pulse width the buffer was overflowing because we
couldn't transfer the data quick enough to avoid the buffer
filling up with samples.

By changing serial output to batch 24 samples and only send
1 byte per sample we are able to continously run without
the buffer overflow!

The sample is also shifted right by 2 in order to get it to
fit into 1 byte. In case there are large gaps we will need
to encode long pulses with a multibyte scheme.

But looks like this approach works!!