Attempts at decoding the disk flux
First issue is that we aren't getting enough data for a complete
revolution. The sum of the bit transitions is 136.5588ms instead of
166.7ms which is nominal.

So could be losing samples? Which would imply the buffer overflow
detection not workingsome how.

The other quirk seems to be that we're getting two bitcell values: one
at 3.3uS and another at 5uS - if the encoding is MFM there should be a
third at 6.7uS?? So are we dealing with FM coding or do we have a bug
dropping the longer pulses (that could explain missing sample time?)