Attempt to improve buffer overflow detection
The theory is that is the flaf complete of complete ISR aren't quick
enough to process the data before the buffer is overwritten then the
half/complete interrupt clears will be set and we can check for this
condition.

If they are then data loss is likely and we need to service the
interrupts quicker or increase the buffer size to gives us more time.